requirements.txt
langbot-plugin
httpx
pyyaml